remove unused uniswapV3Quoter state variable

+++ contracts/libraries/QuoterLib.sol
@@ -1,15 +1,16 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
 pragma solidity ^0.8.0;
 
-import '@uniswap/v3-core/contracts/libraries/SafeCast.sol';
-import '@uniswap/v3-core/contracts/libraries/Simulate.sol';
-import '@uniswap/v3-core/contracts/libraries/TickMath.sol';
-import '@uniswap/v3-core/contracts/interfaces/IUniswapV3Pool.sol';
-import '@uniswap/v3-core/contracts/interfaces/callback/IUniswapV3SwapCallback.sol';
+import { SafeCast } from '@uniswap/v3-core/contracts/libraries/SafeCast.sol';
+import { TickMath } from '@uniswap/v3-core/contracts/libraries/TickMath.sol';
+import { IUniswapV3Pool } from '@uniswap/v3-core/contracts/interfaces/IUniswapV3Pool.sol';
+import { IUniswapV3SwapCallback } from '@uniswap/v3-core/contracts/interfaces/callback/IUniswapV3SwapCallback.sol';
 
-import '@uniswap/v3-periphery/contracts/libraries/Path.sol';
-import '@uniswap/v3-periphery/contracts/libraries/PoolAddress.sol';
-import '@uniswap/v3-periphery/contracts/libraries/CallbackValidation.sol';
+import { Path } from '@uniswap/v3-periphery/contracts/libraries/Path.sol';
+import { PoolAddress } from '@uniswap/v3-periphery/contracts/libraries/PoolAddress.sol';
+import { CallbackValidation } from '@uniswap/v3-periphery/contracts/libraries/CallbackValidation.sol';
+
+import { Simulate } from './Simulate.sol';
 
 /// @title Provides quotes for swaps
 /// @notice Allows getting the expected amount out or amount in for a given swap without executing the swap
@@ -22,14 +23,19 @@ library QuoterLib {
 
     address constant factory = 0x1F98431c8aD98523631AE4a59f267346ea31F984;
 
-    function _getPool(address tokenA, address tokenB, uint24 fee) internal pure returns (IUniswapV3Pool) {
+    function _getPool(
+        address tokenA,
+        address tokenB,
+        uint24 fee
+    ) internal pure returns (IUniswapV3Pool) {
         return IUniswapV3Pool(PoolAddress.computeAddress(factory, PoolAddress.getPoolKey(tokenA, tokenB, fee)));
     }
 
-    function _decodeFirstPool(
-        bytes memory path,
-        bool exactIn
-    ) internal pure returns (IUniswapV3Pool pool, bool zeroForOne) {
+    function _decodeFirstPool(bytes memory path, bool exactIn)
+        internal
+        pure
+        returns (IUniswapV3Pool pool, bool zeroForOne)
+    {
         (address tokenA, address tokenB, uint24 fee) = path.decodeFirstPool();
         pool = _getPool(tokenA, tokenB, fee);
         zeroForOne = exactIn == (tokenA < tokenB);
@@ -131,7 +137,15 @@ library QuoterLib {
         int256 tokenAmount,
         bytes memory swapPath,
         Simulate.State[] memory swapStates
-    ) internal view returns (int256, int256, Simulate.State[] memory) {
+    )
+        internal
+        view
+        returns (
+            int256,
+            int256,
+            Simulate.State[] memory
+        )
+    {
         if (tokenAmount > 0) {
             // swap wbtc to usdc
             (uint256[] memory otherTokenAmounts, Simulate.State[] memory swapStatesEnd) = _quoteExactInput(
@@ -160,7 +174,7 @@ library QuoterLib {
         bytes memory ethSellPath
     ) internal view returns (int256 usdcAmountInBtcSwap, int256 usdcAmountInEthSwap) {
         // btc swap
-        int ethAmountInBtcSwap;
+        int256 ethAmountInBtcSwap;
         Simulate.State[] memory swapStates;
         (usdcAmountInBtcSwap, ethAmountInBtcSwap, swapStates) = _getQuote(
             btcAmountInBtcSwap,
